Added movement scripts by Michal Mucha
These scripts are for the camera controller and event system. According to Michal, 'The way you get it to work rn is kinda slop and I'll change it, but you put the editor script on an empty game object, and the movement controller on the camera or an object that it's attached to.

You click "add waypoint" on the editor object to create waypoints, you move them around until you have your path, then you manually add the waypoints to the array that the movement controller has. When this is finished you'll just use the button but for now you have to do it manually.'